ml_model/requirements.txt
tensorflow>=2.10.0
flask>=2.0.0
flask-cors
opencv-python
numpy
pillow
scikit-learn
matplotlib
